backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-anthropic==0.1.1
anthropic>=0.17.0
Werkzeug==3.0.1
